refactor(codec): consolidate UTF-16 conversion behind simdutf (spec 044)
Finish the migration started in spec 043. Every production call site
of the legacy hand-rolled Utf16LE* converter (16 sites across 10 source
files) now routes through the simdutf-backed implementation. The
public symbol surface and header path are unchanged from pre-spec-043
(encoding::Utf16LE{Encode,Decode,EncodeDirect,ByteLength} declared in
tds/encoding/utf16.hpp); only the implementation behind them is
simdutf-backed. The hand-rolled converter survives as a private
anonymous-namespace fallback inside utf16.cpp, invoked only on invalid
UTF-8 / UTF-16LE input to preserve the bit-for-bit "skip invalid bytes,
continue" semantics from spec 043 Clarification Q1.

Files changed:
- src/tds/encoding/simdutf_wrappers.{hpp,cpp}: deleted; content folded
  into utf16.{hpp,cpp}
- src/include/tds/encoding/utf16.hpp: replaced with the post-rename
  public surface (signature-identical to pre-spec-043) plus a test-only
  testing::LegacyUtf16LE* re-export gated by #ifdef MSSQL_BENCH_BUILD
- src/tds/encoding/utf16.cpp: replaced with simdutf-backed public
  functions plus the hand-rolled LegacyUtf16LE* in anonymous namespace
- src/tds/tds_packet.cpp, src/tds/tds_protocol.cpp: the two spec-043
  consumers swept back to Utf16LE* names (no Simdutf prefix)
- CMakeLists.txt, Makefile: drop simdutf_wrappers.cpp from source lists;
  add -DMSSQL_BENCH_BUILD to test-login7-encoding flags so the test can
  access testing::LegacyUtf16LE* for the spec-043 bit-identity assertion
- test/cpp/test_login7_encoding.cpp: rewire the bit-identity assertions
  to compare encoding::Utf16LE* (public, simdutf-backed) against
  encoding::testing::LegacyUtf16LE* (private legacy)
- test/sql/copy/copy_to_nvarchar_unicode.test: new SQLLogicTest
  regression that round-trips ASCII / Cyrillic / accented Latin / CJK
  ideograph / emoji surrogate-pair / mixed-Unicode values through
  COPY TO MSSQL (BCP) and asserts byte-identical retrieval

The other 9 source files affected by spec 044 (type_converter,
bcp_row_encoder, bcp_writer, azure_fedauth, fedauth_strategy,
manual_token_strategy, mssql_simple_query, tds_column_metadata,
tds_token_parser) show zero diff against main: the Phase A migration
(Utf16LE* -> SimdutfUtf16LE*) and the Phase B rename
(SimdutfUtf16LE* -> Utf16LE*) cancel out at the call-site level. Only
the implementation behind the public symbol surface changes.

Verification:
- make: green
- make test-login7-encoding: 7/7 sections pass (including the spec-043
  bit-identity assertion against 30 fixtures and the invalid-UTF-8
  fallback assertion)
- make test: 103 cases, 3229 assertions, 20 env-driven skips
- make integration-test: 16 cases, 304 assertions
- Audit: zero `SimdutfUtf16LE` references in src/; simdutf_wrappers.*
  files gone; 18 Utf16LE call sites outside the wrapper TU (16 newly
  routed + 2 pre-existing spec-043 consumers)

Out of scope (deferred to a separate session):
- Codec microbenchmark (test/cpp/bench_utf16.cpp, make bench-utf16)
- End-to-end before/after benchmark on Docker SQL Server (100M rows)
- bench_results.md recorded artifact
- clang-format-14 sweep across modified files
