+++ Makefile
@@ -176,12 +176,11 @@ LOGIN7_TEST_SOURCES := \
     src/tds/tds_packet.cpp \
     src/tds/tds_protocol.cpp \
     src/tds/tds_types.cpp \
-    src/tds/encoding/utf16.cpp \
-    src/tds/encoding/simdutf_wrappers.cpp
+    src/tds/encoding/utf16.cpp
 
 LOGIN7_TEST_VCPKG_INSTALLED := build/debug/vcpkg_installed
 LOGIN7_TEST_VCPKG_TRIPLET := $(shell ls $(LOGIN7_TEST_VCPKG_INSTALLED) 2>/dev/null | head -n 1)
-LOGIN7_TEST_FLAGS := -std=c++17 -pthread -Wno-deprecated-declarations
+LOGIN7_TEST_FLAGS := -std=c++17 -pthread -Wno-deprecated-declarations -DMSSQL_BENCH_BUILD
 LOGIN7_TEST_INCLUDES := -I src/include -I duckdb/src/include \
     -I $(LOGIN7_TEST_VCPKG_INSTALLED)/$(LOGIN7_TEST_VCPKG_TRIPLET)/include
 LOGIN7_TEST_LIBS := -L $(LOGIN7_TEST_VCPKG_INSTALLED)/$(LOGIN7_TEST_VCPKG_TRIPLET)/debug/lib -lsimdutf
