+++ src/tds/encoding/utf16.cpp
@@ -1,19 +1,49 @@
+//===----------------------------------------------------------------------===//
+//                         DuckDB MSSQL Extension
+//
+// utf16.cpp
+//
+// UTF-8 <-> UTF-16LE conversion primitives, simdutf-backed.
+//
+// Spec 044 (Codec Layer Consolidation) folded what spec 043 introduced as
+// src/tds/encoding/simdutf_wrappers.{hpp,cpp} back into the original
+// utf16.{hpp,cpp} file path. The public function names are unchanged from
+// the pre-spec-043 legacy converter; only the implementation behind them
+// is now simdutf-backed. The hand-rolled implementation that used to be
+// public survives here as private LegacyUtf16LE* helpers in an anonymous
+// namespace, invoked only by the public functions' invalid-input fallback
+// path (spec 043 Clarification Q1).
+//
+// Invalid-input contract:
+//   1. Pre-validate via simdutf::validate_utf8 (encode direction) or
+//      validate_utf16le (decode direction).
+//   2. Valid input -> SIMD fast path via convert_valid_*.
+//   3. Invalid input -> private LegacyUtf16LE* fallback, preserving the
+//      pre-spec-043 "skip invalid bytes, continue" semantics bit-for-bit.
+//   4. Never throws on invalid input.
+//===----------------------------------------------------------------------===//
+
 #include "tds/encoding/utf16.hpp"
+
+#include <simdutf.h>
+
+#include <cstring>
 
 namespace duckdb {
 namespace tds {
 namespace encoding {
 
 //===----------------------------------------------------------------------===//
-// Fast ASCII check - returns true if string is pure ASCII
+// Private legacy hand-rolled converter (anonymous namespace).
+// Used only as the invalid-input fallback by the public simdutf-backed
+// functions below. Bit-identical to the pre-spec-043 implementation that
+// lived at this same file path.
 //===----------------------------------------------------------------------===//
 
-static inline bool IsAsciiString(const char *data, size_t len) {
-	// Process 8 bytes at a time for better performance
-	size_t i = 0;
+namespace {
 
-	// Check 8 bytes at a time using uint64_t
-	// High bit set in any byte means non-ASCII
+inline bool IsAsciiString(const char *data, size_t len) {
+	size_t i = 0;
 	const uint64_t *ptr64 = reinterpret_cast<const uint64_t *>(data);
 	const size_t chunks = len / 8;
 	for (size_t c = 0; c < chunks; c++) {
@@ -22,8 +52,6 @@ static inline bool IsAsciiString(const char *data, size_t len) {
 		}
 	}
 	i = chunks * 8;
-
-	// Check remaining bytes
 	for (; i < len; i++) {
 		if (static_cast<uint8_t>(data[i]) & 0x80) {
 			return false;
@@ -32,16 +60,9 @@ static inline bool IsAsciiString(const char *data, size_t len) {
 	return true;
 }
 
-//===----------------------------------------------------------------------===//
-// Fast ASCII to UTF-16LE conversion (no decoding needed)
-//===----------------------------------------------------------------------===//
-
-static inline void AsciiToUtf16LE(const char *input, size_t len, std::vector<uint8_t> &result) {
-	// For ASCII, each byte becomes 2 bytes (byte + 0x00)
+inline void AsciiToUtf16LE(const char *input, size_t len, std::vector<uint8_t> &result) {
 	result.resize(len * 2);
 	uint8_t *out = result.data();
-
-	// Unroll loop for better performance
 	size_t i = 0;
 	for (; i + 4 <= len; i += 4) {
 		out[i * 2 + 0] = static_cast<uint8_t>(input[i + 0]);
@@ -53,46 +74,33 @@ static inline void AsciiToUtf16LE(const char *input, size_t len, std::vector<uin
 		out[i * 2 + 6] = static_cast<uint8_t>(input[i + 3]);
 		out[i * 2 + 7] = 0;
 	}
-
-	// Handle remaining bytes
 	for (; i < len; i++) {
 		out[i * 2 + 0] = static_cast<uint8_t>(input[i]);
 		out[i * 2 + 1] = 0;
 	}
 }
 
-//===----------------------------------------------------------------------===//
-// General UTF-8 to UTF-16LE conversion (handles all Unicode)
-//===----------------------------------------------------------------------===//
-
-static void Utf8ToUtf16LEGeneral(const char *input, size_t input_len, std::vector<uint8_t> &result) {
-	result.reserve(input_len * 2);	// Minimum size for ASCII
-
+void Utf8ToUtf16LEGeneral(const char *input, size_t input_len, std::vector<uint8_t> &result) {
+	result.reserve(input_len * 2);
 	size_t i = 0;
 	while (i < input_len) {
 		uint32_t codepoint = 0;
 		uint8_t byte = static_cast<uint8_t>(input[i]);
-
-		// Decode UTF-8 to Unicode codepoint
 		if ((byte & 0x80) == 0) {
-			// Single byte (ASCII): 0xxxxxxx
 			codepoint = byte;
 			i += 1;
 		} else if ((byte & 0xE0) == 0xC0) {
-			// Two bytes: 110xxxxx 10xxxxxx
 			if (i + 1 >= input_len)
 				break;
 			codepoint = ((byte & 0x1F) << 6) | (static_cast<uint8_t>(input[i + 1]) & 0x3F);
 			i += 2;
 		} else if ((byte & 0xF0) == 0xE0) {
-			// Three bytes: 1110xxxx 10xxxxxx 10xxxxxx
 			if (i + 2 >= input_len)
 				break;
 			codepoint = ((byte & 0x0F) << 12) | ((static_cast<uint8_t>(input[i + 1]) & 0x3F) << 6) |
 						(static_cast<uint8_t>(input[i + 2]) & 0x3F);
 			i += 3;
 		} else if ((byte & 0xF8) == 0xF0) {
-			// Four bytes: 11110xxx 10xxxxxx 10xxxxxx 10xxxxxx
 			if (i + 3 >= input_len)
 				break;
 			codepoint = ((byte & 0x07) << 18) | ((static_cast<uint8_t>(input[i + 1]) & 0x3F) << 12) |
@@ -100,90 +108,60 @@ static void Utf8ToUtf16LEGeneral(const char *input, size_t input_len, std::vecto
 						(static_cast<uint8_t>(input[i + 3]) & 0x3F);
 			i += 4;
 		} else {
-			// Invalid UTF-8 byte, skip it
 			i += 1;
 			continue;
 		}
-
-		// Encode Unicode codepoint to UTF-16LE
 		if (codepoint <= 0xFFFF) {
-			// Basic Multilingual Plane (BMP) - single 16-bit code unit
-			result.push_back(static_cast<uint8_t>(codepoint & 0xFF));		  // Low byte
-			result.push_back(static_cast<uint8_t>((codepoint >> 8) & 0xFF));  // High byte
+			result.push_back(static_cast<uint8_t>(codepoint & 0xFF));
+			result.push_back(static_cast<uint8_t>((codepoint >> 8) & 0xFF));
 		} else if (codepoint <= 0x10FFFF) {
-			// Supplementary planes - surrogate pair
 			codepoint -= 0x10000;
 			uint16_t high_surrogate = 0xD800 + ((codepoint >> 10) & 0x3FF);
 			uint16_t low_surrogate = 0xDC00 + (codepoint & 0x3FF);
-			// High surrogate first (in little-endian)
 			result.push_back(static_cast<uint8_t>(high_surrogate & 0xFF));
 			result.push_back(static_cast<uint8_t>((high_surrogate >> 8) & 0xFF));
-			// Low surrogate second
 			result.push_back(static_cast<uint8_t>(low_surrogate & 0xFF));
 			result.push_back(static_cast<uint8_t>((low_surrogate >> 8) & 0xFF));
 		}
-		// Codepoints > 0x10FFFF are invalid, skip them
 	}
 }
 
-//===----------------------------------------------------------------------===//
-// Main UTF-16LE encoding function with fast ASCII path
-//===----------------------------------------------------------------------===//
-
-std::vector<uint8_t> Utf16LEEncode(const std::string &input) {
+std::vector<uint8_t> LegacyUtf16LEEncode(const std::string &input) {
 	std::vector<uint8_t> result;
-
 	if (input.empty()) {
 		return result;
 	}
-
-	// Fast path for ASCII strings (most common case)
 	if (IsAsciiString(input.data(), input.size())) {
 		AsciiToUtf16LE(input.data(), input.size(), result);
 		return result;
 	}
-
-	// Slow path for non-ASCII (full UTF-8 decoding)
 	Utf8ToUtf16LEGeneral(input.data(), input.size(), result);
 	return result;
 }
 
-std::string Utf16LEDecode(const uint8_t *data, size_t byte_length) {
+std::string LegacyUtf16LEDecode(const uint8_t *data, size_t byte_length) {
 	std::string result;
-	result.reserve(byte_length);  // Rough estimate
-
+	result.reserve(byte_length);
 	size_t i = 0;
 	while (i + 1 < byte_length) {
-		// Read UTF-16LE code unit (little-endian)
 		uint16_t code_unit = static_cast<uint16_t>(data[i]) | (static_cast<uint16_t>(data[i + 1]) << 8);
 		i += 2;
-
 		uint32_t codepoint;
-
-		// Check for surrogate pair
 		if (code_unit >= 0xD800 && code_unit <= 0xDBFF) {
-			// High surrogate - expect low surrogate next
 			if (i + 1 >= byte_length)
 				break;
 			uint16_t low_surrogate = static_cast<uint16_t>(data[i]) | (static_cast<uint16_t>(data[i + 1]) << 8);
 			i += 2;
-
 			if (low_surrogate >= 0xDC00 && low_surrogate <= 0xDFFF) {
-				// Valid surrogate pair
 				codepoint = 0x10000 + ((static_cast<uint32_t>(code_unit - 0xD800) << 10) | (low_surrogate - 0xDC00));
 			} else {
-				// Invalid surrogate pair, use replacement character
 				codepoint = 0xFFFD;
 			}
 		} else if (code_unit >= 0xDC00 && code_unit <= 0xDFFF) {
-			// Unexpected low surrogate, use replacement character
 			codepoint = 0xFFFD;
 		} else {
-			// Regular BMP character
 			codepoint = code_unit;
 		}
-
-		// Encode codepoint to UTF-8
 		if (codepoint <= 0x7F) {
 			result.push_back(static_cast<char>(codepoint));
 		} else if (codepoint <= 0x7FF) {
@@ -200,23 +178,15 @@ std::string Utf16LEDecode(const uint8_t *data, size_t byte_length) {
 			result.push_back(static_cast<char>(0x80 | (codepoint & 0x3F)));
 		}
 	}
-
 	return result;
 }
 
-std::string Utf16LEDecode(const std::vector<uint8_t> &data) {
-	return Utf16LEDecode(data.data(), data.size());
-}
-
-size_t Utf16LEByteLength(const std::string &input) {
+size_t LegacyUtf16LEByteLength(const std::string &input) {
 	size_t byte_count = 0;
-
 	size_t i = 0;
 	while (i < input.size()) {
 		uint8_t byte = static_cast<uint8_t>(input[i]);
 		uint32_t codepoint = 0;
-
-		// Decode UTF-8 to get codepoint
 		if ((byte & 0x80) == 0) {
 			codepoint = byte;
 			i += 1;
@@ -242,30 +212,20 @@ size_t Utf16LEByteLength(const std::string &input) {
 			i += 1;
 			continue;
 		}
-
-		// Count UTF-16 bytes needed
 		if (codepoint <= 0xFFFF) {
-			byte_count += 2;  // Single code unit
+			byte_count += 2;
 		} else if (codepoint <= 0x10FFFF) {
-			byte_count += 4;  // Surrogate pair
+			byte_count += 4;
 		}
 	}
-
 	return byte_count;
 }
 
-//===----------------------------------------------------------------------===//
-// Direct UTF-16LE encoding to output buffer (zero allocation hot path)
-//===----------------------------------------------------------------------===//
-
-size_t Utf16LEEncodeDirect(const char *input, size_t input_len, uint8_t *output) {
+size_t LegacyUtf16LEEncodeDirect(const char *input, size_t input_len, uint8_t *output) {
 	if (input_len == 0) {
 		return 0;
 	}
-
-	// Fast path for ASCII strings
 	if (IsAsciiString(input, input_len)) {
-		// Direct ASCII to UTF-16LE conversion (unrolled)
 		size_t i = 0;
 		for (; i + 4 <= input_len; i += 4) {
 			output[i * 2 + 0] = static_cast<uint8_t>(input[i + 0]);
@@ -283,15 +243,11 @@ size_t Utf16LEEncodeDirect(const char *input, size_t input_len, uint8_t *output)
 		}
 		return input_len * 2;
 	}
-
-	// Slow path: full UTF-8 to UTF-16LE conversion
 	size_t out_pos = 0;
 	size_t i = 0;
 	while (i < input_len) {
 		uint32_t codepoint = 0;
 		uint8_t byte = static_cast<uint8_t>(input[i]);
-
-		// Decode UTF-8
 		if ((byte & 0x80) == 0) {
 			codepoint = byte;
 			i += 1;
@@ -317,8 +273,6 @@ size_t Utf16LEEncodeDirect(const char *input, size_t input_len, uint8_t *output)
 			i += 1;
 			continue;
 		}
-
-		// Encode to UTF-16LE
 		if (codepoint <= 0xFFFF) {
 			output[out_pos++] = static_cast<uint8_t>(codepoint & 0xFF);
 			output[out_pos++] = static_cast<uint8_t>((codepoint >> 8) & 0xFF);
@@ -334,6 +288,151 @@ size_t Utf16LEEncodeDirect(const char *input, size_t input_len, uint8_t *output)
 	}
 	return out_pos;
 }
+
+bool ValidateAlignedUtf16Le(const uint8_t *data, size_t byte_length, std::vector<char16_t> &scratch) {
+	if (byte_length == 0) {
+		return true;
+	}
+	const bool aligned = (reinterpret_cast<uintptr_t>(data) & 0x1u) == 0u;
+	const size_t code_units = byte_length / 2;
+	if (aligned) {
+		return simdutf::validate_utf16le(reinterpret_cast<const char16_t *>(data), code_units);
+	}
+	scratch.resize(code_units);
+	std::memcpy(scratch.data(), data, code_units * 2);
+	return simdutf::validate_utf16le(scratch.data(), code_units);
+}
+
+}  // anonymous namespace
+
+//===----------------------------------------------------------------------===//
+// Public simdutf-backed UTF-16LE conversion primitives.
+//===----------------------------------------------------------------------===//
+
+std::vector<uint8_t> Utf16LEEncode(const std::string &input) {
+	if (input.empty()) {
+		return {};
+	}
+
+	const char *src = input.data();
+	const size_t src_len = input.size();
+
+	if (!simdutf::validate_utf8(src, src_len)) {
+		return LegacyUtf16LEEncode(input);
+	}
+
+	const size_t code_units = simdutf::utf16_length_from_utf8(src, src_len);
+	std::vector<uint8_t> result(code_units * 2);
+	if (code_units == 0) {
+		return result;
+	}
+
+	char16_t *out = reinterpret_cast<char16_t *>(result.data());
+	const size_t written = simdutf::convert_valid_utf8_to_utf16le(src, src_len, out);
+	result.resize(written * 2);
+	return result;
+}
+
+size_t Utf16LEEncodeDirect(const char *input, size_t input_len, uint8_t *output) {
+	if (input_len == 0) {
+		return 0;
+	}
+
+	if (!simdutf::validate_utf8(input, input_len)) {
+		return LegacyUtf16LEEncodeDirect(input, input_len, output);
+	}
+
+	// Defensive guard against unaligned output buffers; in practice every
+	// known call site passes a 2-byte-aligned destination.
+	if ((reinterpret_cast<uintptr_t>(output) & 0x1u) != 0u) {
+		return LegacyUtf16LEEncodeDirect(input, input_len, output);
+	}
+
+	char16_t *out = reinterpret_cast<char16_t *>(output);
+	const size_t written = simdutf::convert_valid_utf8_to_utf16le(input, input_len, out);
+	return written * 2;
+}
+
+size_t Utf16LEByteLength(const std::string &input) {
+	if (input.empty()) {
+		return 0;
+	}
+
+	if (!simdutf::validate_utf8(input.data(), input.size())) {
+		return LegacyUtf16LEByteLength(input);
+	}
+
+	return simdutf::utf16_length_from_utf8(input.data(), input.size()) * 2;
+}
+
+std::string Utf16LEDecode(const uint8_t *data, size_t byte_length) {
+	if (byte_length == 0) {
+		return {};
+	}
+
+	std::vector<char16_t> aligned_scratch;
+	if (!ValidateAlignedUtf16Le(data, byte_length, aligned_scratch)) {
+		return LegacyUtf16LEDecode(data, byte_length);
+	}
+
+	const size_t code_units = byte_length / 2;
+	const char16_t *src;
+	std::vector<char16_t> copy;
+	if ((reinterpret_cast<uintptr_t>(data) & 0x1u) == 0u) {
+		src = reinterpret_cast<const char16_t *>(data);
+	} else if (!aligned_scratch.empty()) {
+		src = aligned_scratch.data();
+	} else {
+		copy.resize(code_units);
+		std::memcpy(copy.data(), data, code_units * 2);
+		src = copy.data();
+	}
+
+	const size_t out_bytes = simdutf::utf8_length_from_utf16le(src, code_units);
+	std::string result(out_bytes, '\0');
+	if (out_bytes == 0) {
+		return result;
+	}
+	const size_t written = simdutf::convert_valid_utf16le_to_utf8(src, code_units, &result[0]);
+	result.resize(written);
+	return result;
+}
+
+std::string Utf16LEDecode(const std::vector<uint8_t> &data) {
+	return Utf16LEDecode(data.data(), data.size());
+}
+
+//===----------------------------------------------------------------------===//
+// Test-only re-export of the private legacy hand-rolled converter.
+// Visible only when MSSQL_BENCH_BUILD is defined at compile time (the
+// microbenchmark target and the LOGIN7 unit test enable it). The production
+// extension is built without MSSQL_BENCH_BUILD, so these symbols are
+// unreachable from production code.
+//===----------------------------------------------------------------------===//
+
+#ifdef MSSQL_BENCH_BUILD
+
+namespace testing {
+
+std::vector<uint8_t> LegacyUtf16LEEncode(const std::string &input) {
+	return ::duckdb::tds::encoding::LegacyUtf16LEEncode(input);
+}
+
+std::string LegacyUtf16LEDecode(const uint8_t *data, size_t byte_length) {
+	return ::duckdb::tds::encoding::LegacyUtf16LEDecode(data, byte_length);
+}
+
+size_t LegacyUtf16LEByteLength(const std::string &input) {
+	return ::duckdb::tds::encoding::LegacyUtf16LEByteLength(input);
+}
+
+size_t LegacyUtf16LEEncodeDirect(const char *input, size_t input_len, uint8_t *output) {
+	return ::duckdb::tds::encoding::LegacyUtf16LEEncodeDirect(input, input_len, output);
+}
+
+}  // namespace testing
+
+#endif  // MSSQL_BENCH_BUILD
 
 }  // namespace encoding
 }  // namespace tds
